19: Improve pagination
