+++ project/invoices/views.py
@@ -1,37 +1,13 @@
-import time
-
-from django.shortcuts import render
-from django.http import HttpResponse, JsonResponse
+from django.http import JsonResponse
 from django.core.serializers import serialize
-from django.template import loader
 import json
 from invoices.models import NumberRowPrefix, NumberRowValue
-from django.views.decorators.csrf import csrf_exempt
 from django.core.exceptions import ValidationError
 from invoices.utils.InputValidator import InputValidator
-from functools import wraps
-from user_sessions.models import Token
 from django.db.utils import IntegrityError
 from invoices.utils.Logger import Logger
-from django.core.paginator import Paginator
-from django.shortcuts import render
-
-# Create your views here.
-
-
-def secured_endpoint(endpoint):
-
-    @wraps(endpoint)
-    def wrapper(request, *args, **kwargs):
-        token = request.headers.get("Authentication-Token")
-        if Token.objects.filter(token=token).count() > 0:
-            return endpoint(request, *args, **kwargs)
-        else:
-            error_message = "Invalid token."
-            Logger.error(__name__, error_message)
-            return JsonResponse({"message": error_message}, status=401)
-
-    return wrapper
+from contacts.shared.views_helper import secured_endpoint, paginate_response
+from contacts.shared.response_factory import ResponseFactory
 
 
 @secured_endpoint
@@ -41,15 +17,8 @@ def index(request):
 
 @secured_endpoint
 def number_row_prefix_list(request):
-    const_per_page = 2
-    object_list = NumberRowPrefix.objects.all()  # your queryset
-    paginator = Paginator(object_list, const_per_page)  # Show 10 objects per page
-
-    page_number = request.GET.get('page')  # e.g., ?page=2
-    page_obj = paginator.get_page(page_number)
-    serialized_page_obj = serialize('python', page_obj)
-
-    return JsonResponse({"data": serialized_page_obj, "count": NumberRowPrefix.objects.count()}, safe=False)
+    page_number = request.GET.get('page')
+    return paginate_response(NumberRowPrefix, page_number)
 
 
 @secured_endpoint
@@ -57,11 +26,12 @@ def number_row_prefix_detail(request, id):
     try:
         number_row_prefix = NumberRowPrefix.objects.get(id=id)
         data = serialize('python', [number_row_prefix])
-        return JsonResponse({"data": data[0]}, safe=False)
+        return ResponseFactory.item(data[0])
     except:
         error_message = "Not found."
         Logger.error(__name__, error_message)
-        return JsonResponse({"message": error_message}, status=404)
+        return ResponseFactory.message(error_message, 404)
+
 
 @secured_endpoint
 def number_row_prefix_create(request):
@@ -87,14 +57,5 @@ def number_row_prefix_create(request):
     return JsonResponse({"message": error_message}, status=405)
 
 
-def my_view(request):
-    const_per_page = 2
-    object_list = NumberRowPrefix.objects.all()  # your queryset
-    paginator = Paginator(object_list, const_per_page)  # Show 10 objects per page
 
-    page_number = request.GET.get('page')  # e.g., ?page=2
-    page_obj = paginator.get_page(page_number)
-    serialized_page_obj = serialize('python', page_obj)
-
-    return JsonResponse({"data": serialized_page_obj, "count": NumberRowPrefix.objects.count()}, safe=False)
